Initialize the new rails application
This is going to be a Rails application where the UI is react based
with the initial views served out of Rails and ongoing communication
will happen over action-cable. As such, we'll eliminate all of the
ASSET PIPELINE related stuff (sprockers/turbolinks/etc) and use a
basic webpack-based application.

We've also eliminated the tests so that we can have rspec based tests
in the future.